djangoAdminTest/requirements.txt
Django==2.2
djangorestframework==3.12.4
PyMySQL==1.0.2
pytz==2021.3
sqlparse==0.4.2
